refactor(core): extract reusable types and functions to nx_core
- Moved types related to project graph, inputs, and nx_json to a reusable crate, `nx_core`.
- Removed redundant `transfer_project_graph` function and refactored imports to use `nx_core`.
- Updated `nx_project_graph` and other crates to depend on `nx_core` for shared functionality.
- Migrated utility functions like `find_matching_projects` to `nx_project_graph/utils`.
- Refactored redundant imports and aligned with consolidated module structure.
