+++ Cargo.toml
@@ -18,8 +18,9 @@ nx_project_graph = { path = "crates/nx_project_graph" }
 nx_cache = { path = "crates/nx_cache" }
 nx_terminal = { path = "crates/nx_terminal" }
 nx_watch = { path = "crates/nx_watch" }
-nx_utils = { path = "crates/nx_utils"}
-nx_logger = { path = "crates/nx_logger"}
+nx_utils = { path = "crates/nx_utils" }
+nx_logger = { path = "crates/nx_logger" }
+nx_core = { path = "crates/nx_core" }
 
 anyhow = "1.0.95"
 colored = "2"
