+++ crates/nx_core/src/types.rs
@@ -1,0 +1,3 @@
+pub mod inputs;
+pub mod nx_json;
+pub mod project_graph;
