+++ crates/nx_project_graph/Cargo.toml
@@ -5,21 +5,9 @@ edition = { workspace = true }
 license = { workspace = true }
 
 [dependencies]
-anyhow = { workspace = true }
-tracing = { workspace = true }
-nx_hasher = { workspace = true }
-nx_walker = { workspace = true }
-nx_glob = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
-thiserror = { workspace = true }
-parking_lot = { workspace = true }
-rayon = { workspace = true }
-swc_common = { workspace = true }
-swc_ecma_parser = { workspace = true }
-swc_ecma_ast = { workspace = true }
+nx_core = { workspace = true }
+hashbrown = {workspace = true}
+anyhow = { workspace = true}
+nx_glob = { workspace =  true}
 
 [dev-dependencies]
-assert_fs = { workspace = true }
-tempfile = { workspace = true }
-swc_ecma_dep_graph = { workspace = true }
