+++ crates/nx_project_graph/src/lib.rs
@@ -1,4 +1,1 @@
-/// Placeholder function for project graph operations
-pub fn build_project_graph() -> anyhow::Result<()> {
-    Ok(())
-}
+pub mod utils;
