+++ packages/nx/src/native/types.rs
@@ -1,9 +1,5 @@
 mod file_data;
-mod inputs;
 mod napi_dashmap;
-mod nx_json;
 
 pub use file_data::FileData;
-pub use inputs::*;
 pub use napi_dashmap::NapiDashMap;
-pub use nx_json::*;
