+++ packages/nx/src/native/utils/mod.rs
@@ -1,8 +1,4 @@
-mod find_matching_projects;
 pub mod path;
-
-pub use find_matching_projects::*;
-
 
 #[cfg_attr(not(target_arch = "wasm32"), path = "atomics/default.rs")]
 #[cfg_attr(target_arch = "wasm32", path = "atomics/wasm.rs")]
