+++ packages/nx/src/native/workspace/context.rs
@@ -1,11 +1,12 @@
+use nx_project_graph::utils::find_project_for_path;
+use nx_project_graph::utils::ProjectRootMappings;
 use std::collections::{HashMap, HashSet};
 use std::ops::Deref;
 use std::path::{Path, PathBuf};
 use std::sync::Arc;
 
-use crate::native::project_graph::utils::{find_project_for_path, ProjectRootMappings};
 use crate::native::types::FileData;
-use crate::native::utils::{path::get_child_files,  NxCondvar, NxMutex};
+use crate::native::utils::{path::get_child_files, NxCondvar, NxMutex};
 use crate::native::workspace::files_archive::{read_files_archive, write_files_archive};
 use crate::native::workspace::files_hashing::{full_files_hash, selective_files_hash};
 use crate::native::workspace::types::{
@@ -14,11 +15,11 @@ use crate::native::workspace::types::{
 use crate::native::workspace::{config_files, types::NxWorkspaceFiles, workspace_files};
 use napi::bindgen_prelude::External;
 use nx_hasher::hash;
+use nx_logger::enable_logger;
+use nx_utils::path::ToNormalizedString;
 use rayon::prelude::*;
 use tracing::{trace, warn};
 use xxhash_rust::xxh3;
-use nx_utils::path::ToNormalizedString;
-use nx_logger::enable_logger;
 
 #[napi]
 pub struct WorkspaceContext {
